requirements.txt
pandas
requests
sqlalchemy
tqdm
python-dotenv
